+++ mods/BrewLAN_Plenae/ExpertCamo/mod_info.lua
@@ -1,14 +1,7 @@
 name = "Expert Camo"
-uid = "2529ea71-93ef-41a6-b552-CAMMO000000020"
-version = 2
+uid = "2529ea71-93ef-41a6-b552-CAMMO000000030"
+version = 3
 copyright = "© 2017–2021 Sean Wheeldon"
 description = "Provides framework support for quickly generated reskins of units in mod folders 'TerrainMeshes' and 'CustomMeshes'."
 author = "Balthazar"
 icon = "/mods/BrewLAN_Plenae/ExpertCamo/Mod_Icon.png"
-selectable = true
-enabled = true
-exclusive = false
-ui_only = false
-conflicts = {  }
-before = {  }
-after = {  }